Close block action on enter press in input
